pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.6.5"
description = "CLI-utility to compare files"
authors = ["George Rymarenko <george.rymarenko@icloud.com>"]
packages = [
  { include = "gendiff" },
]
classifiers = [
  "Topic :: Utilities",
]
repository = "https://github.com/georg-remer/python-project-lvl2"

[tool.poetry.dependencies]
python = "^3.6"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
wemake-python-styleguide = "^0.14.1"
pytest-cov = "^2.10.1"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
